refactor: 두 금액을 비교해 마진률을 구하는 메소드를 LottoDrawingResult에서 Money로 이동
