+++ src/test/kotlin/lotto/domain/model/LottoDrawingResultTest.kt
@@ -8,44 +8,7 @@ import org.junit.jupiter.api.Test
 class LottoDrawingResultTest {
 
     @Nested
-    @DisplayName("수익률을 계산하는 calculateMarginRate 메서드 테스트")
-    inner class CalculateMarginRateTest {
-
-        @Test
-        fun `구입 금액이 5천원이고 수익 금액이 5만원일 때 수익률 10을 반환한다`() {
-            val drawingResult = LottoDrawingResult(
-                mapOf(
-                    Rank.FIRST to 0,
-                    Rank.SECOND to 0,
-                    Rank.THIRD to 0,
-                    Rank.FOURTH to 1,
-                    Rank.FIFTH to 0
-                )
-            )
-            val purchaseAmount = Money(5000)
-            val actual = drawingResult.calculateMarginRate(purchaseAmount)
-            assertThat(actual).isEqualTo(Margin(10.0))
-        }
-
-        @Test
-        fun `구입 금액이 14000원이고 수익 금액이 5천원일 때 수익률 0,35714285714285715를 반환한다`() {
-            val drawingResult = LottoDrawingResult(
-                mapOf(
-                    Rank.FIRST to 0,
-                    Rank.SECOND to 0,
-                    Rank.THIRD to 0,
-                    Rank.FOURTH to 0,
-                    Rank.FIFTH to 1
-                )
-            )
-            val purchaseAmount = Money(14000)
-            val actual = drawingResult.calculateMarginRate(purchaseAmount)
-            assertThat(actual).isEqualTo(Margin(0.35714285714285715))
-        }
-    }
-
-    @Nested
-    @DisplayName("총 상금을 계산하는 calculateTotalPrize 메서드 테스트")
+    @DisplayName("추첨 결과로 총 상금을 계산하는 calculateTotalPrize 메서드 테스트")
     inner class CalculateTotalPrizeTest {
 
         @Test
